+++ src/routes.js
@@ -5,6 +5,7 @@
 import {
   Add,
   AddonsControlpanel,
+  Aliases,
   ChangePassword,
   ContactForm,
   Contents,
@@ -31,6 +32,7 @@ import {
   Search,
   Sharing,
   Sitemap,
+  AliasesControlpanel,
   UsersControlpanel,
   UserGroupMembershipControlPanel,
   GroupsControlpanel,
@@ -139,6 +141,10 @@ export const defaultRoutes = [
     component: DatabaseInformation,
   },
   {
+    path: '/controlpanel/aliases',
+    component: AliasesControlpanel,
+  },
+  {
     path: '/controlpanel/moderate-comments',
     component: ModerateComments,
   },
@@ -195,6 +201,10 @@ export const defaultRoutes = [
     component: Sharing,
   },
   {
+    path: '/**/aliases',
+    component: Aliases,
+  },
+  {
     path: '/**/delete',
     component: Delete,
   },
@@ -209,10 +219,6 @@ export const defaultRoutes = [
   {
     path: '/**/history',
     component: History,
-  },
-  {
-    path: '/**/sharing',
-    component: Sharing,
   },
   {
     path: '/**/manage-translations',