+++ packages/generator-volto/CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Bugfix
 
 - Add alias to ESlint config in project generator @sneridagh
+- Add missing `moduleNameMapper`s to default app `package.json` jest config @sneridagh
 
 ### Internal