+++ packages/generator-volto/CHANGELOG.md
@@ -12,6 +12,7 @@
 
 - Add alias to ESlint config in project generator @sneridagh
 - Add missing `moduleNameMapper`s to default app `package.json` jest config @sneridagh
+- Fix storybook for Razzle 4 @sneridagh
 
 ### Internal