+++ packages/generator-volto/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Change Log
 
+## 5.9.1 (unreleased)
+
+### Breaking
+
+### Feature
+
+### Bugfix
+
+### Internal
+
 ## 5.9.0 (2022-09-28)
 
 ### Feature